requirements.txt
Appium-Python-Client==3.1.1
opencv-python==4.8.0.74
pytest==7.4.0
selenium==4.16.0
urllib3==2.1.0 